+++ extensions/gather-v2/src/toggle-camera.ts
@@ -1,0 +1,8 @@
+import { showHUD } from "@raycast/api";
+import { sendGatherKeystroke } from "./utils";
+
+export default async function Command() {
+  if (await sendGatherKeystroke("v", ["shift down", "command down"])) {
+    await showHUD("Toggled camera");
+  }
+}